ingestion/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
idna==3.10
kafka==1.3.5
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
tenacity==9.0.0
twelvedata==1.2.24
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
avro==1.12.0